Inventario de materiales:
-Funciona  nuevo item(verificado).
-Funciona  retirarItem(verificado).
-Funciona  agregarItem(verificado).
-Funciona  Modificar(verificado).
-Funciona  Eliminar(verificado).

-Salir sin implementar(creado Action event).

******Correccion vista table view********
******Cambio a tipo double de cantidad y retiro********
******Textfield solo admite numeros******

+++ src/cl/loccet/controller/InventarioMaterialController.java
@@ -45,11 +45,11 @@ public class InventarioMaterialController extends Controller {
         model.nuevoItem(material);
         cargarDatos();
     }
-    public void agregarMaterial(String idMaterial,int cantidad){
+    public void agregarMaterial(String idMaterial,double cantidad){
         model.agregarMaterial(idMaterial,cantidad);
         cargarDatos();
     }
-    public void retirarMaterial(String idMaterial,int cantidad){
+    public void retirarMaterial(String idMaterial,double cantidad){
         model.retirarMaterial(idMaterial,cantidad);
         cargarDatos();
     }
